+++ CHANGES.md
@@ -3,6 +3,7 @@
 * git-like modules bundled into imaplet.commands library
 * sync up with Irmin 0.10.0
 * integrated gitl as another storage type
+* handle net/disk io in async way in append
 
 ### 0.1.15